Fix subcommand options in qless-js CLI
commander.js distinguishes flags from arguments with values by using the
'--option <value>' syntax.

Also, I made it so the array options (tags and depends) will not default
to [] -- instead they will not be defined at all. This mostly serves to
not show "(default: )" on the help screen.

Also, "tag" needed to be called "tags" to work with qless's `put`
function.

Tested by enqueuing jobs with a certain klass and tags.